app/requirements.txt
psycopg2-binary==2.9.10
redis==5.2.0
schedule==1.2.2